Update .gitignore to include __pycache__ and refine todos.txt item descriptions; refactor main.py to utilize functions for todo operations
