+++ files/todos.txt
@@ -1,8 +1,9 @@
 Listen to Heavy Metal music
-Brush my teeth
+Brush all of my teeth
 Paint the wall
 Drive the car
 Do the weekly shopping
 Hoover the carpet
 Clean the pint glasses
 Clean the caravan
+Work on the graphical interface
